requirements.txt
polymarket-apis>=0.1.0
web3>=6.0.0
eth-account>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
websocket-client>=1.6.0
backoff>=2.2.1
python-telegram-bot>=20.0

